Moved some UART stuff into rp-hal-common.
