+++ rp-hal-common/src/uart/mod.rs
@@ -1,0 +1,6 @@
+//! Shared code and types for Raspberry Pi Silicon UARTS
+
+pub mod common_configs;
+mod utils;
+
+pub use utils::*;
