+++ rp2040-hal/src/uart/utils.rs
@@ -3,7 +3,9 @@ use crate::pac::{uart0::RegisterBlock, UART0, UART1};
 use crate::resets::SubsystemReset;
 use crate::typelevel::Sealed;
 use core::ops::Deref;
-use fugit::HertzU32;
+
+#[doc(inline)]
+pub use rp_hal_common::uart::{DataBits, Parity, StopBits, UartConfig};
 
 /// Error type for UART operations.
 #[derive(Debug)]
@@ -68,79 +70,6 @@ impl Sealed for Enabled {}
 impl State for Disabled {}
 impl Sealed for Disabled {}
 
-/// Data bits
-pub enum DataBits {
-    /// 5 bits
-    Five,
-    /// 6 bits
-    Six,
-    /// 7 bits
-    Seven,
-    /// 8 bits
-    Eight,
-}
-
-/// Stop bits
-pub enum StopBits {
-    /// 1 bit
-    One,
-    /// 2 bits
-    Two,
-}
-
-/// Parity
-/// The "none" state of parity is represented with the Option type (None).
-pub enum Parity {
-    /// Odd parity
-    Odd,
-    /// Even parity
-    Even,
-}
-
-/// A struct holding the configuration for an UART device.
-///
-/// The `Default` implementation implements the following values:
-/// ```ignore
-/// # // can't actually create this with the non_exhaustive attribute
-/// UartConfig {
-///    baudrate: Baud(115_200),
-///    data_bits: DataBits::Eight,
-///    stop_bits: StopBits::One,
-///    parity: None,
-///}
-/// ```
-#[non_exhaustive]
-pub struct UartConfig {
-    /// The baudrate the uart will run at.
-    pub baudrate: HertzU32,
-
-    /// The amount of data bits the uart should be configured to.
-    pub data_bits: DataBits,
-
-    /// The amount of stop bits the uart should be configured to.
-    pub stop_bits: StopBits,
-
-    /// The parity that this uart should have
-    pub parity: Option<Parity>,
-}
-
-impl UartConfig {
-    /// Create a new instance of UartConfig
-    pub const fn new(
-        baudrate: HertzU32,
-        data_bits: DataBits,
-        parity: Option<Parity>,
-        stop_bits: StopBits,
-    ) -> UartConfig {
-        UartConfig {
-            baudrate,
-            data_bits,
-            stop_bits,
-            parity,
-        }
-    }
-}
-
 /// Rx/Tx FIFO Watermark
 ///
 /// Determine the FIFO level that trigger DMA/Interrupt
@@ -161,15 +90,4 @@ pub enum FifoWatermark {
     Bytes24,
     /// Trigger when 28 bytes are (Rx: filled / Tx: available)
     Bytes28,
-}
-
-impl Default for UartConfig {
-    fn default() -> Self {
-        Self {
-            baudrate: HertzU32::from_raw(115_200),
-            data_bits: DataBits::Eight,
-            stop_bits: StopBits::One,
-            parity: None,
-        }
-    }
 }
